+++ packages/core/include/combo.h
@@ -43,6 +43,7 @@
 # include <combo/interface.h>
 # include <combo/common/ocarina.h>
 # include <combo/common/color.h>
+# include <combo/common/cosmetics.h>
 # include <combo/common/actors/En_Tubo_Trap.h>
 
 # if defined(GAME_OOT)