+++ packages/core/include/combo/common/cosmetics.h
@@ -1,0 +1,8 @@
+#ifndef COMBO_COSMETICS_H
+#define COMBO_COSMETICS_H
+
+#include <combo/types.h>
+
+extern u8 gNightBgm;
+
+#endif