+++ packages/core/src/mm/actors/En/En_Suttari.c
@@ -1,6 +1,6 @@
 #include <combo.h>
 
-int Sakon_ShouldEndThiefEscape(Actor_EnSuttari* sakon, PlayState* ctxt) {
+int Sakon_ShouldEndThiefEscape(EnSuttari* sakon, PlayState* ctxt) {
     /* Check if Sakon has dropped the luggage (with speedup) or has finished the escape sequence. */
-    return (sakon->runningState == 6);
+    return (sakon->animIndex == 6);
 }