+++ packages/core/src/mm/scene.c
@@ -1,5 +1,7 @@
 #include <combo.h>
 #include <combo/custom.h>
+#include <combo/common/scene.h>
+#include <combo/common/cosmetics.h>
 
 static EntranceTableEntry defaultEntrance = {
     0, /* Southern Swamp (Clear) */
@@ -165,3 +167,25 @@ void Object_AfterInitContext(void)
 {
     UpdateGameplayKeepForAdult(&gPlay->objectCtx);
 }
+
+u8 gNightBgm;
+EXPORT_SYMBOL(NIGHT_BGM, gNightBgm);
+
+void Scene_CommandSoundSettings(PlayState* play, SceneCmd* cmd) {
+    u8 ambienceId;
+
+    ambienceId = cmd->soundSettings.ambienceId;
+    if(gNightBgm)
+        ambienceId = 0x13;
+
+    play->sceneSequences.seqId = cmd->soundSettings.seqId;
+    play->sceneSequences.ambienceId = ambienceId;
+
+    if (gSaveContext.seqId == (u8)NA_BGM_DISABLED ||
+        // Should be AudioSeq_GetActiveSeqId(u8), but apparently we've defined it as Audio_GetActiveSeqId
+        Audio_GetActiveSeqId(SEQ_PLAYER_BGM_MAIN) == NA_BGM_FINAL_HOURS) {
+        Audio_SetSpec(cmd->soundSettings.specId);
+    }
+}
+
+PATCH_FUNC(0x801303e0, Scene_CommandSoundSettings);